fix: review fixes — extract NoteCard, add save error toasts, deduplicate constants
- Extract NoteCard.svelte from duplicated logic in RawRequirementsModal and quick-ideas page
- Add save error toast via useToasts in raw requirements context (reorder context setup)
- Deduplicate PRD_LABEL: add to SPECIAL_LABELS, import in ForestView
- Fix TrashIcon → CircleCheckIcon for toggle-processed action
- Add bounds check in startEditing for stale index safety
- Remove dead newNoteTextareaElement in quick-ideas page
- Add i18n keys for forest toggle and toggle-processed
- Add /quick-ideas to warmup lists
- Fix unstable {#each} key
- Add concurrent load guard in raw requirements context
