+++ src/lib/components/RawRequirementsModal.svelte
@@ -1,15 +1,12 @@
 <script lang="ts">
 	import * as m from '$lib/paraglide/messages.js';
-	import PencilIcon from '@lucide/svelte/icons/pencil';
-	import CheckIcon from '@lucide/svelte/icons/check';
-	import XIcon from '@lucide/svelte/icons/x';
 	import SparklesIcon from '@lucide/svelte/icons/sparkles';
 	import ExternalLinkIcon from '@lucide/svelte/icons/external-link';
 	import * as Dialog from '$lib/components/ui/dialog/index.js';
 	import * as Tooltip from '$lib/components/ui/tooltip/index.js';
 	import { Button } from '$lib/components/ui/button/index.js';
 	import { Textarea } from '$lib/components/ui/textarea/index.js';
-	import { Badge } from '$lib/components/ui/badge/index.js';
+	import NoteCard from '$lib/components/NoteCard.svelte';
 	import { useRawRequirements } from '$lib/modules/raw-requirements/index.js';
 	import { resolve } from '$app/paths';
 
@@ -18,7 +15,6 @@
 	const rawRequirementsCtx = useRawRequirements();
 
 	let newNoteContent = $state('');
-	let editContent = $state('');
 	let newNoteTextareaElement = $state<HTMLTextAreaElement | null>(null);
 	let notesContainerElement = $state<HTMLDivElement | null>(null);
 
@@ -54,31 +50,6 @@
 			}
 		});
 	}
-
-	function startEditing(index: number) {
-		rawRequirementsCtx.editingIndex = index;
-		editContent = rawRequirementsCtx.notes[index].content;
-	}
-
-	function cancelEditing() {
-		rawRequirementsCtx.editingIndex = null;
-		editContent = '';
-	}
-
-	async function saveEdit(index: number) {
-		await rawRequirementsCtx.updateNote(index, editContent);
-		editContent = '';
-	}
-
-	function handleEditKeydown(event: KeyboardEvent, index: number) {
-		if (event.key === 'Enter' && !event.shiftKey) {
-			event.preventDefault();
-			void saveEdit(index);
-		} else if (event.key === 'Escape') {
-			event.preventDefault();
-			cancelEditing();
-		}
-	}
 </script>
 
 <Dialog.Root open={rawRequirementsCtx.open} onOpenChange={handleDialogOpenChange}>
@@ -93,63 +64,7 @@
 			{:else}
 				<div class="flex flex-col gap-3">
 					{#each rawRequirementsCtx.notes as note, index (note.timestamp + index)}
-						<div class="group rounded-lg border border-border p-3">
-							<div class="mb-1 flex items-center justify-between">
-								<span class="text-xs text-muted-foreground">
-									{note.timestamp}
-								</span>
-								<div class="flex items-center gap-1">
-									{#if note.processed}
-										<Badge
-											variant="default"
-											class="text-[length:var(--text-2xs)]"
-										>
-											{m.raw_requirements_processed()}
-										</Badge>
-									{/if}
-									{#if rawRequirementsCtx.editingIndex !== index}
-										<Button
-											variant="ghost"
-											size="icon-sm"
-											class="opacity-0 group-hover:opacity-100 transition-opacity"
-											onclick={() => startEditing(index)}
-										>
-											<PencilIcon />
-											<span class="sr-only">{m.raw_requirements_edit()}</span>
-										</Button>
-									{/if}
-								</div>
-							</div>
-
-							{#if rawRequirementsCtx.editingIndex === index}
-								<Textarea
-									class="min-h-[60px] text-sm"
-									value={editContent}
-									oninput={(event) => {
-										editContent = event.currentTarget.value;
-									}}
-									onkeydown={(event) => handleEditKeydown(event, index)}
-								/>
-								<div class="mt-2 flex justify-end gap-1">
-									<Button variant="ghost" size="sm" onclick={cancelEditing}>
-										<XIcon />
-										{m.raw_requirements_cancel_edit()}
-									</Button>
-									<Button
-										variant="primary"
-										size="sm"
-										onclick={() => void saveEdit(index)}
-									>
-										<CheckIcon />
-										{m.raw_requirements_save_edit()}
-									</Button>
-								</div>
-							{:else}
-								<p class="whitespace-pre-wrap text-sm text-muted-foreground">
-									{note.content}
-								</p>
-							{/if}
-						</div>
+						<NoteCard {note} {index} />
 					{/each}
 				</div>
 			{/if}
