+++ src/routes/+layout.svelte
@@ -45,9 +45,9 @@
 	setActionsContext();
 	const shortcutsCtx = setKeyboardShortcutsContext();
 	const commandPaletteCtx = setCommandPaletteContext();
+	const toastsCtx = setToastsContext();
 	const rawRequirementsCtx = setRawRequirementsContext();
 	setCreationWizardContext();
-	const toastsCtx = setToastsContext();
 
 	let editingDashboard = $state<Dashboard | null>(null);
 
@@ -59,6 +59,7 @@
 		void preloadCode(resolve('/overview'));
 		void preloadCode(resolve('/sessions'));
 		void preloadCode(resolve('/settings'));
+		void preloadCode(resolve('/quick-ideas'));
 		void shortcutsCtx.loadCustomBindings();
 
 		shortcutsCtx.registerShortcut({
